Team, Player Model Relationship
I am unsure as to how to cascade a change through the players table such that when a team is deleted and they have their foreign key aimed at it, it changes their foreign key to something else (ideally we'd just void the foreign key to represent that they aren't part of a team or aim it at a team defining players that aren't on a roster)

Regardless, I am going to hop onto Mordhau now. Null Byte, if you would, I'd like to talk to you later about how to do that.